+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.5] - 2023-03-27
+
+### Fixed
+
+- Follow ESM import standards ([#78](https://github.com/ciscoheat/sveltekit-superforms/issues/78))
+
 ## [0.6.4] - 2023-03-26
 
 ### Added